Blog post on matching nodes and specifiers (#28)
Added a new blog post on matching nodes and specifiers, with detailed comments and minor edits by @hlapp and other team members.

Because of how we were managing this repository earlier, this PR also includes intermediate changes to other pages, including adding WOP 2017 paper to the products and an arXiv identifier to the tree modeling paper, as well as a previous blog post about the second face-to-face meeting as well updates to author.yml, personnel.yml and Guanyang_Zhang-avatar.jpeg. All of those changes were separately squash-merged into master, and so this PR will not change them in any way, and these changes should be invisible after this squash-merge.